requirements.txt
aiofiles==0.5.0
alembic==1.7.7
anaconda==0.0.1.1
aniso8601==7.0.0
anyio==3.6.2
asttokens==2.0.8
async-exit-stack==1.0.1
async-generator==1.10
backcall==0.2.0
bcrypt==3.2.0
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.1.1
click==7.1.2
cryptography==35.0.0
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.6
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.3.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.0
fastapi==0.68.1
filelock==3.8.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.2
h11==0.13.0
httptools==0.1.2
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.3
jupyter-core==4.11.1
jupyter_client==7.3.5
Mako==1.2.3
MarkupSafe==2.1.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.3
openpyxl==3.0.10
orjson==3.8.2
packaging==21.3
pandas==1.5.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2022.11.23
platformdirs==2.5.2
promise==2.3
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
Pygments==2.13.0
pymongo==4.2.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.4
PyYAML==5.4.1
pyzmq==24.0.1
requests==2.28.1
rsa==4.8
Rx==1.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.25
stack-data==0.5.1
starlette==0.14.2
tornado==6.2
traitlets==5.4.0
typing_extensions==4.1.1
ujson==4.3.0
urllib3==1.26.13
uvicorn==0.13.4
uvloop==0.17.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
watchgod==0.8.2
wcwidth==0.2.5
websockets==8.1
